ai_service/render-requirements.txt
flask
gunicorn
numpy
opencv-contrib-python-headless
requests
networkx
Pillow
